Add files via upload
This is the game that almost everyone played in their childhood. The ROCK PAPER SCISSOR game. In this the rock is winner over the scissor and the paper is winner over the rock and the scissor is winner over the paper. In this the user is played with the computer and computer choose randomly any of them . The game played of best five and at end of game the score and winner will be announced.