Add files via upload
In this project we take url from he respective video which we have to download and past it into the entry box and click on the download and a message is shown after the download is complete that your video download is complete and paste where your this file is stored. All these is done in tkinter.